requirements.txt
beautifulsoup4~=4.12.2
requests~=2.31.0
Pillow~=10.1.0
pytesseract~=0.3.8
cairosvg~=2.7.1
pdfminer.six~=20221105
mammoth~=1.6.0
numpy~=1.26.2
opencv-python~=4.8.1.78
pytest~=7.4.3
pyyaml~=6.0



